+++ appData/src/gb/include/ui.h
@@ -62,6 +62,10 @@ extern UBYTE * text_scroll_addr;
 extern UBYTE text_scroll_width, text_scroll_height;
 extern UBYTE text_scroll_fill;
 
+extern UBYTE text_sound_frames, text_sound_ch;
+extern UBYTE text_sound_bank; 
+extern const UBYTE * text_sound_data;
+
 extern const UBYTE ui_time_masks[];
 
 void ui_init() BANKED;